Remove loader optimizations of dyntrace BIFs
Remove the loader transformations that remove calls to the
undocumented dyntrace BIFs when the runtime system is compiled without
support for dynamic tracing. Those BIFs are only used in a few places
in `file` and `file_io_server`. Since the functions they are used in
are quite heavy, this minor optimization is unlikely to be noticeable
in practice. On the other hand, those transformations slightly slow
down loading of all modules, so removing them is beneficial.