ssh: Check the user name when doing pubkey auth
This name (option {user,UserName}) has nothing to do with the
user name in the OS, but Codenomicum Defensics complains about
not testing it, so therfor thte test is added.
It is off by default (for compatibility), but could be enabled
with the option pk_check_user set to true.