ssl: Enhance handling of certificate chains
TLS-1.3 RFC says that we should handle certificate chains that
have extraneous certs as well as chains that are incomplete but
can be reconstructed or unordered chains.

In this handling we also need to consider partial_chains that may make
an incomplete chain considered complete or should remove some certificates
from the path validation process.

This rewrite of the certificate path handling is done
to consider all these aspects. And it will also improve partial chains
handling to possible shorten a certificate chain from the start and not
only if the full chain fails, as partial chain should not consider certs
above the partial chain anchor cert.