[snmp|agent] Sending a v1 trap on a IPv6 transport caused a crash
Sending a v1 trap on a IPv6 transport is not allowed,
but needlessly caused a crash if was attempted.

In a multi process agent this would cuase the worker
process to crash, but since they are restarted its not
a major issue. But if not a multi process agent, then the
master agent itself would crash, and this *is* a major
issue.

OTP-16920