Check that send and recv data does not overflow ssize_t
send and recv returns the number of sent and received bytes
in an ssize_t, but the amount of data to send is given
in a size_t, or an iov containing size_t:s, so it is possible
to specify more data than send and recv can return.

Therefore we can not send nor recv more data than fits
in a ssize_t or else we can not tell from the return
value of send or recv if all data was sent or if the
recv buffer was filled.