diffable: Fix bug in rejoin_atoms/1
The beam-assembly dump parser in diffable does not parse atoms
containing spaces correctly, leading to, for example, the atom 'foo
bar' to be considered to be two operands. Diffable compensates for
this parser particularity, when important for the correct functioning
of diffable, by using rejoin_atoms/1 to merge split atoms in the
operand list.

This patch fixes a bug in rejoin_atoms/1 where the attempted merging
would stop as soon as a non-atom operand was found.