Implement NSS-style keylog in ssl:connection_information/2
This commit provides the ability to retreive keylog information that
can be used by tools (e.g. Wireshark) to decrypt TLS connections.

The information returned by ssl:connection_information(S, [keylog]) is
a list of keylog items that can be written to a file or otherwise
communicated to a tool understanding keylog.

For TLSv1.3 connections, the initial ssl:listen or ssl:connect call
must include the option {keylog, true}

Reference:
https://developer.mozilla.org/en-US/docs/Mozilla/Projects/NSS/Key_Log_Format