crypto: Suppress valgrind errors at NIF loading
At NIF load crypto tries each of the EC curves to check which that are supported.
If one curve is not supported, there is an error and everyting is deallocated - ctx, key etc.
But, there is an error emitted and by including <openssl/err.h> one could remove the errors with
ERR_* functions. However, including this file triggers compiler errors on Windows.

Since only about 1k lost and only once at the initial loading of crypto, this valgrind error is
suppressed instead.