[kernel] Add new option 'servfail_retry_timeout'
As of version NNN of the DNS server, there is a servfail cache
which has the servfail-ttl default value of 1 sec.
This means if the agent is "to eager" after a servfail, it will
just get the cache'd value (that is, no dns lookup is attempted).
To avoid the situation, a new config option has been added;
servfail_retry_timeout (with default 1500 = 1.5 sec), which
defines the timeout after each failed query, before attempting
a new query.

OTP-16956