Skip literals on v3_core and sys_core_alias
When compiling code with large amount of literals,
there was a performance degradation between Erlang/OTP
22.3 and 23.1.

For example, the gisted code below takes 10 extra seconds:

https://gist.github.com/josevalim/694c1799143fcf25e43aa27e3e11e4c1

In particular, the v3_core pass went from 1.8s to 3.2s.
When profiling the pass, we could see a lot of time was
spent on the new lowering pass in v3_core. In particular,
cerl_trees:mapfold/3 would show up in profilers:

    cerl_trees:mapfold/4   3220377    19.14   2447684  [0.76]

Which is quite curious as the linked code does not trigger
any of the lowering cases.

Under further analysis, this happened because cerl_trees would
traverse literals and other constructs, which is a waste given
there is nothing to lower in many of them.

This commit speeds up both v3_core since cerl_trees is invoked
33% less frequently and the pass is overall 20% faster.
sys_core_alias also got 33% faster. The same technique may be
applicable to other passes.