Explicitly pass in rpo as arguments
Many functions in beam_ssa would receive labels
and compute the rpo from labels. This lead to
multiple calls to rpo which can be inneficient
in some scenarios.

This commit changes these functions to receive
the rpo as argument and shares the RPO in the most
trivial cases found so far. For example, the
`beam_ssa_bool` pass has been made 15-20% faster
when compiling some large modules.

beam_ssa:fold_po/2 has been removed as it is not
used anywhere.