[snmp|agent] Adjust the mib server cache gc
The GC option gclimit where set (to 100) so that it was possible
to cause the cache to grow continuously if the load of the
app was sufficient.
This has now been changed to infinity, to ensure that the cache
is keept as small as possible by default.
Also when at infinity, we now use:

     * ets:select_delete/2

instead of

	ets:select/3 + iterate of the resulting list of keys and
	apply ets:delete/2 to each of them.

Also, make it possible (for *one* process) to subscribe to
(mib server) cache gc events.
This is (mostly) intended for testing.

OTP-16989