macOS: Remove version check in configure script
There is a version check in one of the configure scripts that
tests the version of macOS to ensure that the executables that
are built can be run on the same computer that built them.

The reason for this test has been lost in the mists of time.
Presumably there was a good reason for the test when it was
first added.

Since the test recently has become more complicated because of the
introduction of macOS 11.0 and 11.0.1, and is incorrect for some
versions of macOS, remove it. If there turns out that the version test
would be useful, it can be added back later (but in a less buggy way,
preferably using `awk` instead of `sed` to make the version parsing
simpler and more accurate).