Remove warning text about the `--` operation from documentation
The `--` operation was optimized in Erlang/OTP 22 so that its worst
case complexity is O(N*log(N)), where N is the total size of the input
lists. Therefore, the warning in the documentation saying that the
time complexity is proportional to length(A)*length(B) is incorrect
and is no longer needed. Note that Erlang/OTP 21 will no longer be
supported when Erlang/OTP 24 gets released.