Make the Erlang installation directory relocatable
Update the erl and start scripts to find and set ROOTDIR dynamically when
needed, removing the dependency on a preset hardcoded absolute ROOTDIR path.
This makes the 2 scripts independent of their actual location and allows the
Erlang installation directory to be moved freely on Unix-like systems.

Keep the usage and configuration of FINAL_ROOTDIR as-is at the moment for
backwards compatibility.