logger: Make logger_simple print pretty reports
When the simple logger has been started we are late enough
in the boot cycle that we can load any modules that we need
in order to handle printing the crash. We wrap the call to
logger_formatter in a try catch just to make sure that in case
we should fail we actually get something printed.