stdlib: Document epp option 'location'
Note that option 'scan_opts' is removed (it was not used). Although
its introduction might seem like a good idea, the implementation is a
bit complicated. In particular macro expansion needs to be updated.

If epp is called from the compiler, scanner option 'text' is the only
relevant one since neither spaces nor comments are handled by the
erl_parse module.