beam_ssa_dead: Lower the threshold for giving up
There is a mechanism for giving up a search for a forced label
in beam_ssa_dead. Based on the compilation time for this code:

https://gist.github.com/josevalim/2c175678a0f85568eb001dcb97024ef5

where the running time for beam_ssa_dead still dominates, it seems
prudent to lower the threshold for when the seach should be given
up.

https://bugs.erlang.org/browse/ERL-1302