shell_docs: Clear ansi at new-line
We should clear the ansi sequence whenever there is a newline
so that tools that print things at newline are not broken.

This fix cheats a bit and only does the newline ansi check when
printing blocks of characters as we know that ansi codes can
only happen within inline elments and never inbetween blocks.