Use maps:from_list/1 on maps:intersect_with/3
The current implementation is not efficient when
the maps being compared are mostly distinct.
It happens that collecting all keys and building
the map once at the end is the most efficient
way of implementing intersection.

Benchmarks show this approch is often 4x faster in
the mostly distinct cases, otherwise it behaves
within 5-10% of the current implementation. The only
case it is considerably slower is when working with
small sets of integers (~10 elements) but those
are done in nanoseconds anyway.

Benchmaking code and results:
https://gist.github.com/josevalim/759f3a9c2613119d3e38cd4224840d32