[snmp|agent] Introduce notification specific worker
In order to make it possible to serialize notifications
'through' the agent, its now possible to configure the
agent to use a worker process dedicated to handle
notifications.

Note that this does not remove the issue with inform-requests,
which require a reply (which may not arrive). A retransmission
may still cause reorder.

OTP-17022