Don't crash if the number of file descriptors is unlimited
On macOS 11, the maximum number of file descriptors can be
unlimited.  `erl` refuses to start if it can't read out the
limit.

It is not practical for `erl` to use a huge number such as 134217728
as the default value, because that value is also used as the size of
the port table. (Making the port table's size dynamic is not something
we want do in a maintenance release.)

Therefore we will set it to 24576, which is the value returned by
`sysctl kern.maxfilesperproc`, and which seems to be the practical
limit.  (Raising that limit might be possible, but it can only be done
by disabling System Integrity Protection (SIP) and editing a bunch of
configuration files.  Anyone who has done that will have to set ulimit
explicitly to number of file descriptors they want to use.)

https://bugs.erlang.org/browse/ERL-1417