erl_lint: Remove -behaviour() checking
`-behaviour()` requires that the behavior module is built before
the module we're compiling, and that it can be found in the build
system's (!!) code path.

This is very annoying to deal with so we've decided to remove it
as `dialyzer` already does this check, and in a far better fashion
than the linter.

The only thing `dialyzer` lacks is that it doesn't understand the
`behavior_info/1` callback (and raises a warning to that effect),
but we consider that less important than fixing this issue.