ssl: Implement trial decryption of early data
If early data is not enabled on the server, it will ignore the
early_data extension and then will skip past early data by attempting to
deprotect received records using the handshake traffic key, discarding
records which fail deprotection, up to the configured
max_early_data_size. Once a record is deprotected successfully, it is
treated as the start of the client's second flight and the server
proceeds as with an ordinary 1-RTT handshake.