Add erl option for reading config from file descriptor
Make it possible for users to specify the option "-configfd FD" when
executing the erl command. When this option is given the system will
try to read and parse configuration parameters from the file
descriptor. At the moment the only configuration format that is
supported is the format that is supported when using the -config
option.

Motivation
----------

Allow users to configure an Erlang system without writing files to
disk (which can be an issue in some container environments with
read-only file systems).