erts: Fix vector count for bit strings
When a bit string was preceeded by another term it would
incorrectly not increment the number of vectors needed to
contain the term during pending connect.

The testcase has been fixed to contain wrapped un-aligned
bit-strings and also to not sent the data across the
distribution port before the test as that destroys all
un-aligned bit-strings.