Add the new operator ^ for pinning pattern variables
This allows you to annotate already-bound pattern variables as ^X, like
in Elixir. This is less error prone when code is being refactored and
moved around so that variables previously new in a pattern may become
bound, or vice versa, and makes it easier for the reader to see the intent
of the code. An new compiler flag 'warn_unpinned_vars' (disabled by default)
provides warnings for all uses of already bound variables in patterns that
have not been marked with ^.