shell history: allow configurable history module
In addition to 'enabled' or 'disabled' setting, kernel
shell_history variable may be an atom that allows to use
any module as callback module for shell history.
This enabled to fetch shell history from different sources,
for example, cloud-based environment, or other nodes of
a cluster.
This commit also fixes test case that was using -kernel
shell_history 'true' instead of 'enabled'.