ssl: Correct cert path construction, solves CVE-2020-35733
The certificate path used as input to
public_key:pkix_path_validation/3 is basically the reverse
certificate chain sent by the peer, excluding the ROOT that may
or may not be sent by the peer. TLS-1.3 does however specify that
that chains could be unordered (as log as the peer certificate is
first) and chains can contain extraneous certificates. In
addition certificates can be missing, but if we own them so that we
can reconstruct the chain that should also be acceptable too and
also the user could choose to put their trust in an intermediate
certificate, that is shortening the chain that should be validated.

When completing the support for all of the above in commit
d24a220c3b867caef83026ba31d2656366da4322 a bug was introduced that
could be exploited to construct fake certificate chains that could
be validated. This commit fixes that. Also comments
and name enhancments are done to make the code easier to understand.
In doing so limit chain alternatives for avoiding using more memory
than necessary. A better solution will be added in later commit.