gen: clarify usage of 'self' as parent PID
The atom `self` is used instead of parent PID for processes without a
parent, like in the case of monitor or link. The use of the atom `self`
could be mistaken for a typo.

This patch changes the atom declaration to use single quotes to clarify
the intended use of `'self'` as an atom and hopefully removes the
possibility that `'self'` is interpreted as a typo when read by a fellow
developer.