YCF: Zero variables declared inside yielding function in debug mode
YCF's debug mode detected a lot of false positives (pointers allocated
on the function stack for a yielding function that points to the
function stack) because uninitialized variables sometimes happens to
contain addresses to the stack. This commit fixes this issue by
automatically writing zeros to these variables when debug mode is
activated.