Trap `maps:from_keys/2`
The function `maps:from_keys/2` did not trap (yield) before this
commit which could lead to unfair scheduling. This commit fixes this
issue by making sure that `maps:from_keys/2` calls with a long list as
parameter value traps.

Trapping versions for all major functions that the implementation of
`maps:from_keys/2` makes use of are generated with the Yielding C Fun*
tool.

* $ERL_TOP/erts/lib_src/yielding_c_fun/README.md