wxNotificationMessage support
This enables native os notification events on the platform. For implementing messaging and background apps this is especially important. E.g. imagine an app wanting to announce 'Update available' or 'New Mail Received' or 'Download Completed'.

For MacOS even which discourages applications raising themself to focus, using the native OS notifications is the only correct way to get the users attention to an app.

Note that the underlying wxWidgets implementation in 3.0.x is very basic and has been completely reworked in 3.1.x for best experience use wxWidgets 3.1