os: Eliminate wrappers for BIFs
For convenience during development of c2d70945dce several environment
manipulation functions gained wrappers; that is, the documented function
(such as `os:putenv/2`) is an Erlang function that calls the BIF doing
the real work. Eliminate alls these wrapper functions (for example,
`os:putenv/2` will now be a BIF implemented in C).

We are doing this preliminary refactoring to facilitate adding extended
error information to the BIFs in the `os` module.