jit: Optimize three-way select_val
Optimize the case when a select_val has exactly two choices which
branch to the same label or a fail destination and where the values
only differ by one bit.

The optimization makes use of the observation that (V == X || V == Y)
is equivalent to (V | (X ^ Y)) == (X | Y) when (X ^ Y) has only one
bit set, which allows the select_val to be implemented with only one
compare. This optimization is unconditionally performed by both GCC
and LLVM on x86_64.

When this optimization is implemented in the code generator, the
change in performance compared to the unmodified system varies
depending on the probability distribution of the value the select_val
operates on. For illustration, consider a {select_val, X, Fail,
[{1,Dest}, {2, Dest}]}-instruction where the input only consists of
ones. In this case a slowdown of 8% is apparent. If the same
instruction is fed only twos, the slowdown decreases to around 3%.  If
the input is changed to be a 1/3 each of 1, 2, and 3, the performance
change depends strongly on how well the branch predictor functions in
the unmodified system. If the input is a repeated sequence i.e. (123)+
or 1+2+3+ the slowdown is between 3% and 4%. If the input is
completely random, in which case the branch predictor cannot help
things, the optimization provides a speedup of ~25%.

For the larger benchmarks in ebench's `small` class, the only
statistically significant performance changes are that `decode` is ~4%
faster and `fib` is ~2.5% faster.