erts_qsort: Use qsort from the C standard library for small arrays
This commit makes erts_qsort sort with qsort from the C standard
library when it needs to sort small arrays. Small arrays are also
sorted with qsort in recursive calls so this affect input arrays of
all sizes.

This optimization improves the performance of maps:from_list/1 and
maps:from_keys/2.

Thanks to Sverker Eriksson (@sverker) for helping me in finding this
optimization.