Make the type tag lookup more robust
In order to gather all info about a type that EEP-48 requires,
we refer to EDoc #tag{} records.
These records do not contain the type name at the top level,
so this lookup was done just by line number.

When a header file is included into another file,
it might happen that the position information of two distinct type
definitions is the same - they are still different types,
but the lookup by line will return the wrong one.

This lead to the following error (on kernel/src/disk_log.erl):

    name: dlog_size
    arity: 0
    type attr: {attribute,82,type,{file_error,{type,82,term,[]},[]}}

    edoc: error in layout 'edoc_layout_chunks': {'EXIT',
                                           {{badmatch,{file_error,[]}},
                                            [{edoc_layout_chunks,meta_type_sig,4,
                                              [{file,"edoc_layout_chunks.erl"},
                                               {line,170}]},

This change introduces a more sophisticated and robust lookup,
which fixes the problem of types defined in include files.