Make sure that math:{acos,asin}/1 consistently fail for domain errors
We require a C compiler that supports C99, but on Solaris it could
happen that even though the C compiler (gcc) supports C99, the runtime
libraries could still behave as a previous version (X/Open or C90).

Specifically, the acos() and asin() function in C90 return 0.0 on a
domain error, but sets the errno to EDOM. To ensure that we raise an
exception, we will need to check errno after calling the those
functions.

While at it, also provide extended BIF information for `math:acosh/1`
and `math:atanh/1`.