shell: read records from module in escript/archive
Previously the `rr/1` shell function failed with the below error when
called for a module which was loaded from an archive or escript archive,
because beam_lib:chunks uses regular `file:open/2` to read from the
file. Now `erl_prim_loader:get_file/1` is used which supports file paths
in archives.

```
{error, beam_lib, {file_error,"<path_to_archive>/test_app.ez/test_app/ebin/test.beam", enotdir}}
```