Make `maps:keys/1` and `maps:values/1` trapping
The functions `maps:keys/1` and `maps:values/1` did not trap (yield)
before this commit which could lead to unfair scheduling. This commit
fixes this issue by making sure that `maps:keys/1` and `maps:values/1`
calls with a big map as parameter value traps.

Trapping versions for all major functions that the implementation of
the functions make use of are generated with the Yielding C Fun* tool.

* $ERL_TOP/erts/lib_src/yielding_c_fun/README.md